Fix perlin noise world generation and modularize app
