+++ ProjectState.md
@@ -3,7 +3,7 @@
 ## Finished
 - Basic voxel world rendering with free camera controls.
 - Title screen with adjustable world size, Start Game and Exit buttons.
-- Perlin noise terrain generation on game start.
+- Perlin noise terrain generation on game start with corrected frequency for varied height.
 
 ## WIP
 - None
