+++ src/state.rs
@@ -1,0 +1,8 @@
+use bevy::prelude::*;
+
+#[derive(States, Default, Debug, Clone, Eq, PartialEq, Hash)]
+pub enum AppState {
+    #[default]
+    Menu,
+    Playing,
+}
